+++ examples/plugins/altemplate/src/template.rs
@@ -30,7 +30,7 @@ use suricata::applayer::{
     state_get_tx_iterator, AppLayerEvent, AppLayerTxData, State, Transaction,
     APP_LAYER_PARSER_EOF_TC, APP_LAYER_PARSER_EOF_TS, APP_LAYER_PARSER_OPT_ACCEPT_GAPS,
 };
-use suricata::applayer::{AppLayerResultRust, StreamSliceRust};
+use suricata_ffi::applayer::{AppLayerResultRust, StreamSliceRust};
 use suricata_ffi::conf::conf_get;
 use suricata_ffi::{
     build_slice, cast_pointer, export_state_data_get, export_tx_data_get, SCLogError, SCLogNotice,
